requirements.txt
redshift-connector==2.0.918
python-dotenv==1.0.0
boto3==1.29.0
pandas==2.1.4
pyarrow==14.0.2
pyspark==3.5.0